+++ test/stepwise_tests/step_6_tests.dart
@@ -1,6 +1,5 @@
-import 'package:flutter_test/flutter_test.dart';
-
 import 'package:http/http.dart' show ClientException;
+import 'package:test/test.dart';
 
 import 'package:nordigen_integration/nordigen_integration.dart';