+++ test/utilities/create_random_requisition.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
 
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 
 import 'package:nordigen_integration/nordigen_integration.dart';